Connect the app to remote AWS RDS Postresql database and S3 bucket for media storage
1)Update the settings file to handle both local and production environments.
Add the option to store the remote AWS environments variables in a .env file which will not be part of the git repository.
The settings file is first loading the environment file if exists and populate the variables.
NOTE: it's the responsibility of the prod team to setup an environment file with all required variables and run the docker with that file

2)default_storage is used to interact with file storage or S3 storage under the hood
Refactor the cleanup db and populate db scripts to use default_storage rather than os system library.
boto3 client is used to perform tasks related to S3